🐛 Fix(중복 엔티티 발생): 중복 엔티티 발생 버그 수정
- JpaRepository 연관 수정
- JpaAuditing 중복 어노테이션 제거
- BoardUser 엔티티 중복 제거
- User 엔티티 권한 제거
