+++ src/main/java/com/wolf/workflow/WorkFlowApplication.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
-@EnableJpaAuditing
 public class WorkFlowApplication {
 
     public static void main(String[] args) {
