+++ src/main/java/com/wolf/workflow/board/entity/BoardUser.java
@@ -1,37 +1,42 @@
 package com.wolf.workflow.board.entity;
 
-import com.wolf.workflow.common.Timestamped;
-import com.wolf.workflow.user.entity.BoardUserRole;
-import com.wolf.workflow.user.entity.Participation;
 import com.wolf.workflow.user.entity.User;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@Table(name = "board_user_id")
+@Table(name = "board_user")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class BoardUser extends Timestamped {
+public class BoardUser {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "board_user_id")
+    private Long id;
 
-    @Column
-    private Participation participation;
-
-    @Column
-    private BoardUserRole role;
-
-    @ManyToOne(fetch =FetchType.LAZY)
-    @JoinColumn(name = "user_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-    @ManyToOne(fetch =FetchType.LAZY)
-    @JoinColumn(name = "board_id")
-    private Board board;
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Participation participation;
 
-
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private BoardUserRole boardUserRole;
 
 }
