--- src/main/java/com/wolf/workflow/board/entity/enums/Participation.java
@@ -1,5 +1,0 @@
-package com.wolf.workflow.board.entity.enums;
-
-public enum Participation {
-    Disable,Enable
-}
