+++ src/main/java/com/wolf/workflow/user/repository/UserRepository.java
@@ -1,4 +1,8 @@
 package com.wolf.workflow.user.repository;
 
-public interface UserRepository {
+import com.wolf.workflow.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
 }
